backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
qdrant-client>=1.7.1
openai==1.6.1
sqlalchemy==2.0.23
psycopg2-binary>=2.9.10  # Updated for Python 3.13 compatibility
alembic==1.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
